requirements.txt
# API 요청 및 웹 관련
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
bleach==6.1.0
lxml==4.9.3

# 데이터베이스
mysql-connector-python==8.3.0
SQLAlchemy==2.0.23

# 구성 및 환경
python-dotenv==1.0.0
PyYAML==6.0.1

# 데이터 처리
pandas==2.1.1
numpy==1.26.0

# 로깅 및 출력
colorama==0.4.6
tqdm==4.66.1
prettytable==3.9.0

# 파일 및 시스템
python-dateutil==2.8.2
openpyxl==3.1.2

# 유틸리티
retry==0.9.2
tenacity==8.2.3

# HTML 보고서 생성
jinja2==3.1.2
markdown==3.5

# 병렬 처리
aiohttp==3.9.1

# 테스트 및 개발
pytest==7.4.3
pytest-mock==3.12.0
